refactor code, extract owl animation into seperate kotlin file, fresh start screen
