+++ app/src/main/java/com/example/learningapp/ui/theme/Color.kt
@@ -18,7 +18,6 @@ val Orange50 = Color(0x59FA8231)
 val Purple40 = Color(0xFF6650a4)
 val PurpleGrey40 = Color(0xFF625b71)
 val Pink40 = Color(0xFF7D5260)
-val Green40 = Color(0xFF56BD0E)
 val grey = Color(0xFF5A6361)
 
 // light mode colors
